refactor: Clean up imports of `borer.test` and example a bit more
